[ONNX] Wrap op validation inputs and add fx_utils.py
[ghstack-poisoned]
