+++ torch/onnx/_internal/fx/exporter.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import copy
-import dataclasses
 import functools
 import inspect
 import itertools
@@ -14,11 +13,8 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 import onnx
 import onnxscript  # type: ignore[import]
-from onnxscript import evaluator, opset18  # type: ignore[import]
-from onnxscript.function_libs.torch_aten import (  # type: ignore[import]
-    graph_building,
-    ops,
-)
+from onnxscript import evaluator  # type: ignore[import]
+from onnxscript.function_libs.torch_aten import graph_building  # type: ignore[import]
 
 import torch
 import torch._C
@@ -33,143 +29,12 @@ from torch.nn.utils import stateless
 from torch.onnx import _constants, _type_utils
 
 from torch.onnx._internal import _beartype
-from torch.onnx._internal.fx import diagnostics
+from torch.onnx._internal.fx import diagnostics, fx_utils
 from torch.utils import _pytree
 
 
 # TODO: Separate into individual components.
 # TODO: make_fx lose stack info https://github.com/pytorch/pytorch/issues/90276
-
-
-TORCH_ONNX_OPSET = onnxscript.values.Opset(domain="torch.onnx", version=1)
-
-
-@onnxscript.script(opset=TORCH_ONNX_OPSET)
-def prims_convert_element_type(tensor, dtype: int):
-    return opset18.Cast(tensor, to=dtype)
-
-
-@onnxscript.script(opset=TORCH_ONNX_OPSET)
-def aten_getitem(self, i):
-    # TODO(justinchuby): Support
-    # i = opset18.Unsqueeze(i, opset18.Constant(value_ints=[0]))
-    # return opset18.Gather(self, i, axis=0)
-    return opset18.SequenceAt(self, i)
-
-
-# A simple lookup table for atenlib functions
-_ATENLIB_FUNCTIONS = {
-    "getitem": aten_getitem,
-    "prims::convert_element_type": prims_convert_element_type,
-    "aten::abs": ops.core.aten_abs,
-    "aten::acos": ops.core.aten_acos,
-    "aten::acosh": ops.core.aten_acosh,
-    "aten::add": ops.core.aten_add,
-    "aten::addmm": ops.core.aten_addmm,
-    "aten::amax": ops.core.aten_amax,
-    "aten::amin": ops.core.aten_amin,
-    "aten::arange": ops.core.aten_arange_start,
-    "aten::asin": ops.core.aten_asin,
-    "aten::asinh": ops.core.aten_asinh,
-    "aten::atan": ops.core.aten_atan,
-    "aten::atanh": ops.core.aten_atanh,
-    "aten::bmm": ops.core.aten_bmm,
-    "aten::ceil": ops.core.aten_ceil,
-    "aten::clamp_max": ops.core.aten_clamp_max,
-    "aten::clamp_min": ops.core.aten_clamp_min,
-    "aten::clamp": ops.core.aten_clamp,
-    "aten::clone": ops.core.aten_clone,
-    "aten::convolution": ops.core.aten_convolution,
-    "aten::cos": ops.core.aten_cos,
-    "aten::cosh": ops.core.aten_cosh,
-    "aten::detach": ops.core.aten_detach,
-    "aten::div": ops.core.aten_div,
-    "aten::dot": ops.core.aten_dot,
-    "aten::empty": ops.core.aten_empty,
-    "aten::empty_like": ops.core.aten_empty_like,
-    "aten::eq": ops.core.aten_eq,
-    "aten::equal": ops.core.aten_equal,
-    "aten::exp": ops.core.aten_exp,
-    "aten::exp2": ops.core.aten_exp2,
-    "aten::expand": ops.core.aten_expand,
-    "aten::erf": ops.core.aten_erf,
-    "aten::fmod": ops.core.aten_fmod,
-    "aten::full": ops.core.aten_full,
-    "aten::full_like": ops.core.aten_full_like,
-    "aten::ge": ops.core.aten_ge,
-    "aten::gt": ops.core.aten_gt,
-    "aten::isinf": ops.core.aten_isinf,
-    "aten::log": ops.core.aten_log,
-    "aten::le": ops.core.aten_le,
-    "aten::log10": ops.core.aten_log10,
-    "aten::log1p": ops.core.aten_log1p,
-    "aten::log_softmax": ops.special.aten_special_log_softmax,
-    "aten::log2": ops.core.aten_log2,
-    "aten::logaddexp": ops.core.aten_logaddexp,
-    "aten::logaddexp2": ops.core.aten_logaddexp2,
-    "aten::logcumsumexp": ops.core.aten_logcumsumexp,
-    "aten::logdet": ops.core.aten_logdet,
-    "aten::logsumexp": ops.core.aten_logsumexp,
-    "aten::lt": ops.core.aten_lt,
-    "aten::matmul": ops.core.aten_matmul,
-    "aten::maximum": ops.core.aten_maximum,
-    "aten::minimum": ops.core.aten_minimum,
-    "aten::mm": ops.core.aten_mm,
-    "aten::mul": ops.core.aten_mul,
-    "aten::ne": ops.core.aten_ne,
-    "aten::neg": ops.core.aten_neg,
-    "aten::new_full": ops.core.aten_new_full,
-    "aten::adaptive_avg_pool1d": ops.nn.aten_adaptive_avg_pool1d,
-    "aten::adaptive_avg_pool2d": ops.nn.aten_adaptive_avg_pool2d,
-    "aten::adaptive_avg_pool3d": ops.nn.aten_adaptive_avg_pool3d,
-    "aten::celu": ops.nn.aten_celu,
-    "aten::elu": ops.nn.aten_elu,
-    "aten::embedding": ops.core.aten_embedding,
-    "aten::gelu": ops.nn.aten_gelu,
-    "aten::leaky_relu": ops.nn.aten_leaky_relu,
-    "aten::linear": ops.nn.aten_linear,
-    "aten::logsigmoid": ops.nn.aten_log_sigmoid,
-    "aten::relu": ops.nn.aten_relu,
-    "aten::relu6": ops.nn.aten_relu6,
-    "aten::selu": ops.core.aten_selu,
-    "aten::upsample_nearest2d": ops.nn.aten_upsample_nearest2d,
-    "aten::nonzero": ops.core.aten_nonzero,
-    "aten::ones_like": ops.core.aten_ones_like,
-    "aten::ones": ops.core.aten_ones,
-    "aten::permute": ops.core.aten_permute,
-    "aten::pow": ops.core.aten_pow,
-    "aten::reciprocal": ops.core.aten_reciprocal,
-    "aten::remainder": ops.core.aten_remainder,
-    "aten::repeat": ops.core.aten_repeat,
-    "aten::reshape": ops.core.aten_reshape,
-    "aten::round": ops.core.aten_round,
-    "aten::rsqrt": ops.core.aten_rsqrt,
-    "aten::rsub": ops.core.aten_rsub,
-    "aten::sigmoid": ops.core.aten_sigmoid,
-    "aten::sign": ops.core.aten_sign,
-    "aten::sin": ops.core.aten_sin,
-    "aten::sinh": ops.core.aten_sinh,
-    "aten::slice": ops.core.aten_slice,
-    "aten::softmax": ops.special.aten_special_softmax,
-    "aten::split": ops.core.aten_split,
-    "aten::sqrt": ops.core.aten_sqrt,
-    "aten::sub": ops.core.aten_sub,
-    "aten::t": ops.core.aten_t,
-    "aten::tan": ops.core.aten_tan,
-    "aten::tanh": ops.core.aten_tanh,
-    "aten::topk": ops.core.aten_topk,
-    "aten::unsqueeze": ops.core.aten_unsqueeze,
-    "aten::view": ops.core.aten_view,
-    "aten::where": ops.core.aten_where,
-    "aten::xlogy": ops.special.aten_special_xlogy,
-    "aten::zeros": ops.core.aten_zeros,
-    "aten::zeros_like": ops.core.aten_zeros_like,
-    "aten::native_layer_norm": ops.core.aten_native_layer_norm,
-    "aten::transpose": ops.core.aten_transpose,
-    "aten::sum": ops.core.aten_sum_dim_IntList,
-    "aten::argmin": ops.core.aten_argmin,
-    "aten::argmax": ops.core.aten_argmax,
-}
 
 
 def _onnx_function_diagnose_call_message_formatter(
@@ -200,46 +65,12 @@ _diagnose_onnx_function = diagnostics.diagnose_call(
     diagnostic_message_formatter=_onnx_function_diagnose_call_message_formatter,
     diagnostic_modifier=_onnx_function_diagnose_call_append_symbolic_source_location,
 )
-for key, onnx_function in _ATENLIB_FUNCTIONS.items():
+for key, onnx_function in fx_utils._ATENLIB_FUNCTIONS.items():
     if isinstance(onnx_function, FunctionType):
-        _ATENLIB_FUNCTIONS[key] = _diagnose_onnx_function(onnx_function)
+        fx_utils._ATENLIB_FUNCTIONS[key] = _diagnose_onnx_function(onnx_function)
 onnxscript.OnnxFunction.__call__ = _diagnose_onnx_function(
     onnxscript.OnnxFunction.__call__
 )
-
-
-def _create_op_overload_to_exporter_key_table() -> Dict[
-    Union[torch._ops.OpOverload, Callable], str
-]:
-    # TODO(justinchuby): Improve how the table is constructed.
-    table: Dict[Union[torch._ops.OpOverload, Callable], str] = {}
-
-    for op_namespace in (torch.ops.aten, torch.ops.prims):
-        for attr_name in dir(op_namespace):
-            op_overload_packet = getattr(op_namespace, attr_name)
-
-            if not isinstance(op_overload_packet, torch._ops.OpOverloadPacket):
-                continue
-
-            exporter_look_up_key = op_overload_packet._qualified_op_name
-            if _ATENLIB_FUNCTIONS.get(exporter_look_up_key) is None:
-                # This aten op doesn't have ONNX exporter.
-                continue
-
-            for overload_name in op_overload_packet.overloads():
-                op_overload = getattr(op_overload_packet, overload_name)
-                # This line maps torch.ops.aten.add.Tensor, torch.ops.aten.add.Scalar, torch.ops.aten.add.out, etc
-                # to "aten::add". This means the exporter for "aten::add" is used for all overloads of "aten::add".
-                # This is applied to all ops under torch.ops.aten.
-                #
-                # TODO(wechi): in the future, we might want to write individual exporter for each overload, if,
-                # for example, they have different type promotion rules. If so, just map different overloads to
-                # different exporter keys.
-
-                table[op_overload] = op_overload_packet._qualified_op_name
-    # TODO(justinchuby): is baddbmm different?
-    table[torch.ops.aten.baddbmm.default] = "aten::baddbmm"
-    return table
 
 
 class ModuleExpansionTracer(torch.fx._symbolic_trace.Tracer):
@@ -351,34 +182,6 @@ def _module_expansion_symbolic_trace(
         for name, (_, wrapped) in patched_torch_methods.items():
             # wrapped is the original version of `torch.name`.
             setattr(torch, name, wrapped)
-
-
-# Dictionary that maps torch.ops.aten.* to exporter look up key; e.g.,
-# _OP_OVERLOAD_TO_EXPORTER_KEY_TABLE[torch.add.Tensor] is "aten::add".
-_OP_OVERLOAD_TO_EXPORTER_KEY_TABLE = _create_op_overload_to_exporter_key_table()
-
-
-@_beartype.beartype
-def _create_onnx_friendly_decomposition_table() -> Dict[
-    torch._ops.OpOverload, Callable
-]:
-    decomposition_table: Dict[torch._ops.OpOverload, Callable] = {}
-    for op_overload, decomp_fn in torch._decomp.decomposition_table.items():
-        # Skip decomposition into "prim::*" ops, because they are not generally supported by ONNX.
-        # Skip decomposition for op_overload as long as that op_overload has a corresponding ONNX exporter.
-        if (
-            "torch._refs" in decomp_fn.__module__
-            or op_overload in _OP_OVERLOAD_TO_EXPORTER_KEY_TABLE
-        ):
-            continue
-        decomposition_table[op_overload] = decomp_fn
-    return decomposition_table
-
-
-# This is a subset of PyTorch's built-in aten-to-aten decomposition. If an aten
-# op (e.g., torch.ops.aten.add.Tensor) has exporter, we exclude the op's decomposition
-# function in the _ONNX_FRIENDLY_DECOMPOSITION_TABLE.
-_ONNX_FRIENDLY_DECOMPOSITION_TABLE = _create_onnx_friendly_decomposition_table()
 
 
 def _retrieve_or_adapt_input_to_graph_set(fx_node_arg, fx_name_to_onnxscipt_value):
@@ -500,22 +303,24 @@ def _fill_tensor_meta(
             onnxscript_value.name = name
 
 
-# FIXME(titaiwang): ORT not supports current graph (input type)
+@_beartype.beartype
 def _validate_op_between_ort_torch(
-    node: torch.fx.Node, symbolic_fn, torch_args, torch_kwargs
+    node: torch.fx.Node,
+    symbolic_fn: onnxscript.OnnxFunction,
+    torch_args: tuple,
+    torch_kwargs: dict,
 ):
     """Validate the op between ONNX Runtime and PyTorch."""
     # op-level validation
-    # TODO(titaiwang): Change ORTEvaluator to ReferenceEvaluator
     # Symbolic_fn should have the same output as node.target (torch ops)
     try:
         with evaluator.default_as(evaluator.ort_evaluator):
             expected_outputs = node.target(*torch_args, **torch_kwargs)  # type: ignore[operator]
-            numpy_args = [
-                arg.numpy() if isinstance(arg, torch.Tensor) else arg
-                for arg in torch_args
-            ]
-            ort_outputs = symbolic_fn(*numpy_args, **torch_kwargs)
+            # TODO(titaiwang): Expose _convert_tensor_to_numpy and _convert_kwargs_for_onnx?
+            input_onnx = [fx_utils._convert_tensor_to_numpy(x) for x in torch_args]
+            # deal with dtype and device
+            kwargs_onnx = fx_utils._convert_kwargs_for_onnx(torch_kwargs)
+            ort_outputs = symbolic_fn(*input_onnx, **kwargs_onnx)
 
             for ort_output, expected_output in zip(ort_outputs, expected_outputs):
                 try:
@@ -611,7 +416,7 @@ def _export_fx_node_to_onnxscript(
     ],
     tracer: graph_building.TorchScriptTracingEvaluator,
     fx_module_with_metadata: torch.fx.GraphModule,
-    options: ExportOptions,
+    options: fx_utils.ExportOptions,
 ):
     # Record stack trace of node in diagnostic.
     node_stack_trace = node.stack_trace
@@ -663,13 +468,13 @@ def _export_fx_node_to_onnxscript(
             exporter_key = "getitem"
         elif (
             isinstance(node.target, torch._ops.OpOverload)
-            and node.target in _OP_OVERLOAD_TO_EXPORTER_KEY_TABLE
+            and node.target in fx_utils._OP_OVERLOAD_TO_EXPORTER_KEY_TABLE
         ):
-            exporter_key = _OP_OVERLOAD_TO_EXPORTER_KEY_TABLE[node.target]
+            exporter_key = fx_utils._OP_OVERLOAD_TO_EXPORTER_KEY_TABLE[node.target]
         else:
             raise RuntimeError(f"Unknown call_function target: {node.target}")
         # Only the latest opset version is only supported in atenlib for now
-        symbolic_fn = _ATENLIB_FUNCTIONS.get(exporter_key)
+        symbolic_fn = fx_utils._ATENLIB_FUNCTIONS.get(exporter_key)
         if symbolic_fn is None:
             raise RuntimeError(f"Cannot find function for {exporter_key}")
         # Map FX inputs to ONNX inputs and fill optional inputs with default values.
@@ -748,7 +553,7 @@ def _export_fx_node_to_onnxscript(
 
 @diagnostics.diagnose_call(diagnostics.rules.atenlib_fx_to_onnx)
 def _export_fx_to_onnxscript(
-    fx_module_with_metadata: torch.fx.GraphModule, options: ExportOptions
+    fx_module_with_metadata: torch.fx.GraphModule, options: fx_utils.ExportOptions
 ):
 
     # Initialize the ONNX graph
@@ -859,7 +664,7 @@ def _export(
     **kwargs,
 ) -> Union["onnx.ModelProto", bytes]:
 
-    options = ExportOptions()
+    options = fx_utils.ExportOptions()
     options.update(**kwargs)
     # Apply decomposition table to the input graph.
     # Make sure the feed-in "module" is stateless.
@@ -921,7 +726,7 @@ def export(
         graph_module,
         args,
         opset_version=opset_version,
-        decomposition_table=_ONNX_FRIENDLY_DECOMPOSITION_TABLE,
+        decomposition_table=fx_utils._ONNX_FRIENDLY_DECOMPOSITION_TABLE,
         use_binary_format=use_binary_format,
         op_level_debug=op_level_debug,
     )
@@ -987,7 +792,7 @@ def export_without_kwargs(
         # Function optimized by _dynamo doesn't have None in args.
         tuple(arg for arg in bound_args if arg is not None),
         opset_version=opset_version,
-        decomposition_table=_ONNX_FRIENDLY_DECOMPOSITION_TABLE,
+        decomposition_table=fx_utils._ONNX_FRIENDLY_DECOMPOSITION_TABLE,
         use_binary_format=use_binary_format,
         op_level_debug=op_level_debug,
     )
@@ -1295,30 +1100,6 @@ def save_model_with_external_data(
 
     # model_location should be a pure file name such as "file_name.onnx", not "folder/file_name.onnx".
     onnx.save(onnx_model_with_initializers, os.path.join(basepath, model_location))
-
-
-@dataclasses.dataclass
-class ExportOptions:
-    """Options for FX-ONNX export.
-    Attributes:
-        opset_version: The export ONNX version.
-        use_binary_format: Whether to Return ModelProto in binary format.
-        decomposition_table: The decomposition table for graph ops. Default is for torch ops, including aten and prim.
-    """
-
-    opset_version: int = _constants.ONNX_DEFAULT_OPSET
-    use_binary_format: bool = True
-    op_level_debug: bool = False
-    decomposition_table: Dict[torch._ops.OpOverload, Callable] = dataclasses.field(
-        default_factory=lambda: _ONNX_FRIENDLY_DECOMPOSITION_TABLE
-    )
-
-    def update(self, **kwargs):
-        for key, value in kwargs.items():
-            if hasattr(self, key):
-                setattr(self, key, value)
-            else:
-                warnings.warn(f"ExportOptions has no attribute {key}")
 
 
 # Register a few argument formatter
